+++ lib/grinder_sdk.dart
@@ -42,10 +42,9 @@ class Dart {
   /// Run a dart [script] using [run_lib.run].
   ///
   /// Returns the stdout.
-  static String run(String script,
-      {List<String> arguments : const [], bool quiet: false,
-       String packageRoot, String workingDirectory, int vmNewGenHeapMB,
-       int vmOldGenHeapMB, List<String> vmArgs : const []}) {
+  static String run(String script, {List<String> arguments: const [],
+      bool quiet: false, String packageRoot, RunOptions runOptions,
+      int vmNewGenHeapMB, int vmOldGenHeapMB, List<String> vmArgs: const []}) {
     List<String> args = [];
 
     args.addAll(vmArgs);
@@ -65,8 +64,8 @@ class Dart {
     args.add(script);
     args.addAll(arguments);
 
-    return run_lib.run(_sdkBin('dart'), arguments: args, quiet: quiet,
-        workingDirectory: workingDirectory);
+    return run_lib.run(_sdkBin('dart'),
+        arguments: args, quiet: quiet, runOptions: runOptions);
   }
 
   static String version({bool quiet: false}) {
@@ -96,12 +95,12 @@ class Pub {
    * even if the pubspec.lock file is up-to-date with respect to the
    * pubspec.yaml file.
    */
-  static void get({bool force: false, String workingDirectory}) {
+  static void get({bool force: false, RunOptions runOptions}) {
     FileSet pubspec = new FileSet.fromFile(new File('pubspec.yaml'));
     FileSet publock = new FileSet.fromFile(new File('pubspec.lock'));
 
     if (force || !publock.upToDate(pubspec)) {
-      _run('get', workingDirectory: workingDirectory);
+      _run('get', runOptions: runOptions);
     }
   }
 
@@ -110,13 +109,14 @@ class Pub {
    * even if the pubspec.lock file is up-to-date with respect to the
    * pubspec.yaml file.
    */
-  static Future getAsync({bool force: false, String workingDirectory}) {
+  static Future getAsync({bool force: false, RunAsyncOptions runOptions}) {
     FileSet pubspec = new FileSet.fromFile(new File('pubspec.yaml'));
     FileSet publock = new FileSet.fromFile(new File('pubspec.lock'));
 
     if (force || !publock.upToDate(pubspec)) {
-      return run_lib.runAsync(_sdkBin('pub'), arguments: ['get'],
-          workingDirectory: workingDirectory).then((_) => null);
+      return run_lib
+          .runAsync(_sdkBin('pub'), arguments: ['get'], runOptions: runOptions)
+          .then((_) => null);
     }
 
     return new Future.value();
@@ -125,31 +125,35 @@ class Pub {
   /**
    * Run `pub upgrade` on the current project.
    */
-  static void upgrade({String workingDirectory}) {
-    _run('upgrade', workingDirectory: workingDirectory);
+  static void upgrade({RunOptions runOptions}) {
+    _run('upgrade', runOptions: runOptions);
   }
 
   /**
    * Run `pub upgrade` on the current project.
    */
-  static Future upgradeAsync({String workingDirectory}) {
-    return run_lib.runAsync(_sdkBin('pub'), arguments: ['upgrade'],
-        workingDirectory: workingDirectory).then((_) => null);
+  static Future upgradeAsync({RunAsyncOptions runOptions}) {
+    return run_lib
+        .runAsync(_sdkBin('pub'),
+            arguments: ['upgrade'], runOptions: runOptions)
+        .then((_) => null);
   }
 
   /**
    * Run `pub downgrade` on the current project.
    */
-  static void downgrade({String workingDirectory}) {
-    _run('downgrade', workingDirectory: workingDirectory);
+  static void downgrade({RunOptions runOptions}) {
+    _run('downgrade', runOptions: runOptions);
   }
 
   /**
    * Run `pub downgrade` on the current project.
    */
-  static Future downgradeAsync({String workingDirectory}) {
-    return run_lib.runAsync(_sdkBin('pub'), arguments: ['downgrade'],
-        workingDirectory: workingDirectory).then((_) => null);
+  static Future downgradeAsync({RunAsyncOptions runOptions}) {
+    return run_lib
+        .runAsync(_sdkBin('pub'),
+            arguments: ['downgrade'], runOptions: runOptions)
+        .then((_) => null);
   }
 
   /**
@@ -157,18 +161,14 @@ class Pub {
    *
    * The valid values for [mode] are `release` and `debug`.
    */
-  static void build({
-      String mode,
-      List<String> directories,
-      String workingDirectory,
-      String outputDirectory}) {
+  static void build({String mode, List<String> directories,
+      RunOptions runOptions, String outputDirectory}) {
     List args = ['build'];
     if (mode != null) args.add('--mode=${mode}');
     if (outputDirectory != null) args.add('--output=${outputDirectory}');
     if (directories != null && directories.isNotEmpty) args.addAll(directories);
 
-    run_lib.run(_sdkBin('pub'), arguments: args,
-        workingDirectory: workingDirectory);
+    run_lib.run(_sdkBin('pub'), arguments: args, runOptions: runOptions);
   }
 
   /**
@@ -176,40 +176,38 @@ class Pub {
    *
    * The valid values for [mode] are `release` and `debug`.
    */
-  static Future buildAsync({
-      String mode,
-      List<String> directories,
-      String workingDirectory,
-      String outputDirectory}) {
+  static Future buildAsync({String mode, List<String> directories,
+      RunAsyncOptions runOptions, String outputDirectory}) {
     List args = ['build'];
     if (mode != null) args.add('--mode=${mode}');
     if (outputDirectory != null) args.add('--output=${outputDirectory}');
     if (directories != null && directories.isNotEmpty) args.addAll(directories);
 
-    return run_lib.runAsync(_sdkBin('pub'), arguments: args,
-        workingDirectory: workingDirectory).then((_) => null);
+    return run_lib
+        .runAsync(_sdkBin('pub'), arguments: args, runOptions: runOptions)
+        .then((_) => null);
   }
 
   /// Run `pub run` on the given [package] and [script].
   ///
   /// If [script] is null it defaults to the same value as [package].
   static String run(String package,
-      {List<String> arguments, String workingDirectory, String script}) {
+      {List<String> arguments, RunOptions runOptions, String script}) {
     var scriptArg = script == null ? package : '$package:$script';
     List args = ['run', scriptArg];
     if (arguments != null) args.addAll(arguments);
-    return run_lib.run(_sdkBin('pub'), arguments: args,
-        workingDirectory: workingDirectory);
+    return run_lib.run(_sdkBin('pub'), arguments: args, runOptions: runOptions);
   }
 
-  static String version({bool quiet: false}) => _AppVersion.parse(
-      _run('--version', quiet: quiet)).version;
+  static String version({bool quiet: false}) =>
+      _AppVersion.parse(_run('--version', quiet: quiet)).version;
 
   static PubGlobal get global => _global;
 
-  static String _run(String command, {bool quiet: false, String workingDirectory}) {
-    return run_lib.run(_sdkBin('pub'), quiet: quiet, arguments: [command],
-        workingDirectory: workingDirectory);
+  static String _run(String command,
+      {bool quiet: false, RunOptions runOptions}) {
+    return run_lib.run(_sdkBin('pub'),
+        quiet: quiet, arguments: [command], runOptions: runOptions);
   }
 }
 
@@ -252,7 +250,8 @@ class Dart2js {
     args.add('-o${outFile.path}');
     args.add(sourceFile.path);
 
-    return run_lib.runAsync(_sdkBin('dart2js'), arguments: args)
+    return run_lib
+        .runAsync(_sdkBin('dart2js'), arguments: args)
         .then((_) => null);
   }
 
@@ -270,8 +269,8 @@ class Analyzer {
   /// Analyze a single [File] or path ([String]).
   static void analyze(fileOrPath,
       {Directory packageRoot, bool fatalWarnings: false}) {
-    analyzeFiles([fileOrPath], packageRoot: packageRoot,
-        fatalWarnings: fatalWarnings);
+    analyzeFiles([fileOrPath],
+        packageRoot: packageRoot, fatalWarnings: fatalWarnings);
   }
 
   /// Analyze one or more [File]s or paths ([String]).
@@ -308,7 +307,8 @@ class DartFmt {
   }
 
   static String _run(String option, String target, {bool quiet: false}) =>
-      run_lib.run(_sdkBin('dartfmt'), quiet: quiet, arguments: [option, target]);
+      run_lib.run(_sdkBin('dartfmt'),
+          quiet: quiet, arguments: [option, target]);
 }
 
 /// Access the `pub global` commands.
@@ -320,19 +320,19 @@ class PubGlobal {
   /// Install a new Dart application.
   void activate(String packageName, {bool force: false}) {
     if (force || !isActivated(packageName)) {
-      run_lib.run(_sdkBin('pub'), arguments: ['global', 'activate', packageName]);
+      run_lib.run(_sdkBin('pub'),
+          arguments: ['global', 'activate', packageName]);
       _activatedPackages.add(packageName);
     }
   }
 
   /// Run the given installed Dart application.
   String run(String package,
-      {List<String> arguments, String workingDirectory, String script}) {
+      {List<String> arguments, RunOptions runOptions, String script}) {
     var scriptArg = script == null ? package : '$package:$script';
     List args = ['global', 'run', scriptArg];
     if (arguments != null) args.addAll(arguments);
-    return run_lib.run(_sdkBin('pub'), arguments: args,
-        workingDirectory: workingDirectory);
+    return run_lib.run(_sdkBin('pub'), arguments: args, runOptions: runOptions);
   }
 
   /// Return the list of installed applications.
@@ -342,8 +342,8 @@ class PubGlobal {
     //discoveryapis_generator 0.6.1
     //...
 
-    var stdout = run_lib.run(
-        _sdkBin('pub'), arguments: ['global', 'list'], quiet: true);
+    var stdout =
+        run_lib.run(_sdkBin('pub'), arguments: ['global', 'list'], quiet: true);
 
     var lines = stdout.trim().split('\n');
     return lines.map((line) {
@@ -397,7 +397,7 @@ abstract class PubApp {
   /// If [script] is provided, the sub-script will be run. So
   /// `new PubApp.global('grinder').run(script: 'init');` will run
   /// `grinder:init`.
-  String run(List<String> arguments, {String script, String workingDirectory});
+  String run(List<String> arguments, {String script, RunOptions runOptions});
 
   String toString() => packageName;
 }
@@ -429,8 +429,8 @@ class _AppVersion {
   static _AppVersion parse(String output) {
     var lastSpace = output.lastIndexOf(' ');
     if (lastSpace == -1) return new _AppVersion(output);
-    return new _AppVersion(output.substring(0, lastSpace),
-        output.substring(lastSpace + 1));
+    return new _AppVersion(
+        output.substring(0, lastSpace), output.substring(lastSpace + 1));
   }
 
   String toString() => '$name $version';
@@ -446,13 +446,11 @@ class _PubGlobalApp extends PubApp {
   void activate({bool force: false}) =>
       Pub.global.activate(packageName, force: force);
 
-  String run(List<String> arguments, {String script, String workingDirectory}) {
+  String run(List<String> arguments, {String script, RunOptions runOptions}) {
     activate();
 
     return Pub.global.run(packageName,
-        script: script,
-        arguments: arguments,
-        workingDirectory: workingDirectory);
+        script: script, arguments: arguments, runOptions: runOptions);
   }
 }
 
@@ -464,12 +462,10 @@ class _PubLocalApp extends PubApp {
   // TODO: Implement: call a `Pub.isActivated/Pub.isInstalled`.
   bool get isActivated => throw new UnsupportedError('unimplemented');
 
-  void activate({bool force: false}) { }
+  void activate({bool force: false}) {}
 
-  String run(List<String> arguments, {String script, String workingDirectory}) {
+  String run(List<String> arguments, {String script, RunOptions runOptions}) {
     return Pub.run(packageName,
-        script: script,
-        arguments: arguments,
-        workingDirectory: workingDirectory);
+        script: script, arguments: arguments, runOptions: runOptions);
   }
 }