+++ lib/src/run_utils.dart
@@ -9,7 +9,7 @@ import 'dart:convert';
 import '../grinder.dart';
 
 Stream<String> toLineStream(Stream<List<int>> s) =>
-  s.transform(UTF8.decoder).transform(const LineSplitter());
+    s.transform(UTF8.decoder).transform(const LineSplitter());
 
 logStdout(String line) {
   log(line);